requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
dnspython==2.7.0
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.10.1
pygments==2.19.2
pymongo==4.14.1
python-dotenv==1.1.1
python-multipart==0.0.22
pyyaml==6.0.2
rich==14.1.0
sentry-sdk==2.35.1
shellingham==1.5.4
starlette==0.47.2
typer==0.16.1
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
sqlalchemy==2.0.40
greenlet==3.3.1
alembic==1.12.0
asyncpg==0.31.0
aiosqlite==0.20.0
pytest==8.3.2
pytest-asyncio==0.21.1
pillow==10.0.0
mongomock==4.3.0
